refactor: move alt view interfaces and registry to session-alt-views/
